refactor: Moved PeerTransport logic into PeerStream; refactored functions of PeerStream.
